% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_check_args.R
\name{add_missing_dots}
\alias{add_missing_dots}
\title{Helper to add the missing dots marked argumeyns to the calling_args list with the "missing" value.}
\usage{
add_missing_dots(calling_args, dotted_args)
}
\arguments{
\item{calling_args}{list of calling parameters-arguments couples.}

\item{dotted_args}{Character vector or dotted-marked parameters.}
}
\description{
Helper to add the missing dots marked argumeyns to the calling_args list with the "missing" value.
}
