% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_check_args.R
\name{adjust_args_names}
\alias{adjust_args_names}
\title{Helper to convert the names of args in a vector of "dots" and "non_dots"}
\usage{
adjust_args_names(args)
}
\arguments{
\item{args}{Character vector reporting the calling function arguments to check (parameter names).}
}
\value{
Returns the args vector.
}
\description{
Helper to convert the names of args in a vector of "dots" and "non_dots"
}
