% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{alert_generator.character}
\alias{alert_generator.character}
\title{S3 method of alert_generator for 'character' class}
\usage{
\method{alert_generator}{character}(type, alert_message, eval_env)
}
\arguments{
\item{type}{string equal to one of: 'error', 'message' and 'warning'. Determines the type of alert.}

\item{alert_message}{list of character vector or a single character vector that follow the cli
syntax. If list it is formatted following the scheme explained in format_cli_list.}

\item{eval_env}{numeric indicating the calling frame in which evaluates the alert in respect
to which is called. So it indicates the number of frames to look up in the calling stack.}
}
\description{
S3 method of alert_generator for 'character' class
}
