% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_lists_generators.R
\name{alert_generator}
\alias{alert_generator}
\title{Generic that allows to generate error, warning calls based on the specified type and using custom messages.}
\usage{
alert_generator(type, alert_message)
}
\arguments{
\item{type}{string equal to one of: 'error', 'message' and 'warning'. Determines the type of alert.}

\item{alert_message}{list of character vector or a single character vector that follow the cli
syntax. If list it is formatted following the scheme explained in format_cli_list.}
}
\description{
Generic that allows to generate error, warning calls based on the specified type and using custom messages.
}
