% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{all2}
\alias{all2}
\title{Custom all to signal NULL values in input through a warning and to control
NA values in input and output.}
\usage{
all2(..., na.rm = FALSE, na_res = FALSE)
}
\arguments{
\item{...}{passing arguments to all}

\item{na.rm}{logical indicating if the NA in input must be removed before evaluation}

\item{na_res}{logical indicating the result of all2 if all returns NA in output}
}
\description{
Custom all to signal NULL values in input through a warning and to control
NA values in input and output.
}
