% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{any_data_frame}
\alias{any_data_frame}
\title{Random data frame generator}
\usage{
any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)
}
\arguments{
\item{rows}{Number of rows of the generated data frame.
If \code{rows} is a single number all data frames will have
this number of rows. If \code{rows} is a numeric vector of
length 2 it will produce data frames with rows between
a minimum and maximum, inclusive. For example
\code{rows = c(1L, 10L)} would produce data frames with rows
between 1 and 10. To produce empty tibbles set \code{rows = 0L}
or a range like \code{rows = c(0L, 10L)}.}

\item{cols}{Number of columns of the generated data frame.
If \code{cols} is a single number all data frames will have
this number of columns. If \code{cols} is a numeric vector
of length 2 it will produce data frames with columns
between a minimum and maximum, inclusive. For example
\code{cols = c(1L, 10L)} would produce data frames with
columns between 1 and 10. To produce empty tibbles
set \code{cols = 0L} or a range like \code{cols = c(0L, 10L)}.}

\item{frac_na}{The approximate fraction of \code{NA} values to
be included in the vector. A fraction of 0, the default,
will include no \code{NA} values. A fraction of 0.5 would
include around 50\% \code{NA} values.}
}
\value{
A \code{quickcheck_generator} generator
}
\description{
Generate random data frames.
}
\examples{
any_data_frame() |> show_example()
any_data_frame(rows = 10L) |> show_example()
any_data_frame(cols = 5L, frac_na = 0.5) |> show_example()
}
