% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{any_flat_list}
\alias{any_flat_list}
\title{Flat list generator}
\usage{
any_flat_list(len = c(1L, 10L), frac_na = 0)
}
\arguments{
\item{len}{Length of the generated vectors. If \code{len} is a
single number all vectors will have this length. If
\code{len} is a numeric vector of length 2 it will produce
vectors with lengths between a minimum and maximum,
inclusive. For example \code{len = c(1L, 10L)} would produce
vectors with lengths between 1 and 10. To produce empty vectors
set \code{len = 0L} or a range like \code{len = c(0L, 10L)}.}

\item{frac_na}{The approximate fraction of \code{NA} values to
be included in the vector. A fraction of 0, the default,
will include no \code{NA} values. A fraction of 0.5 would
include around 50\% \code{NA} values.}
}
\value{
A \code{quickcheck_generator} generator
}
\description{
Generate lists where each element is an atomic vector with a length of 1.
}
\examples{
any_flat_list() |> show_example()
any_flat_list(len = 10L, frac_na = 0.5) |> show_example()
}
