% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{any_undefined}
\alias{any_undefined}
\title{Any undefined value generator}
\usage{
any_undefined()
}
\value{
A \code{quickcheck_generator} object.
}
\description{
Generate undefined values. In this case undefined values include \code{NA},
\code{NA_integer_}, \code{NA_real_}, \code{NA_character_}, \code{NA_complex_}, \code{NULL}, \code{-Inf},
\code{Inf}, and \code{NaN}. Values generated are always scalars.
}
\examples{
any_undefined() \%>\% show_example()
}
