% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{anything_defined}
\alias{anything_defined}
\title{Any R object generator with no undefined values}
\usage{
anything_defined()
}
\value{
A \code{quickcheck_generator} generator
}
\description{
Generate any R object which isn't an undefined value or doesn't contain
undefined values. This doesn't actually generate any possible object,
just the most common ones, namely atomic vectors, lists, and tibbles.
None of the objects created will include undefined values like \code{NA}, \code{NULL},
\code{Inf}, \code{NaN}, or be empty vectors or tibbles.
}
\examples{
anything_defined() |> show_example()
}
