% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasversal_internals.R
\name{apply_predicate}
\alias{apply_predicate}
\title{Apply The predicate function on the object and returns the resulting logical vector.}
\usage{
apply_predicate(object, predicate, inverse)
}
\arguments{
\item{object}{Vector or list.}

\item{predicate}{A predicate function, usually an "is.something" function.}

\item{inverse}{Boolean, whether to invert the check direction in the sense that the predicate must be not satisfied for all the elements (default FALSE).}
}
\description{
Apply The predicate function on the object and returns the resulting logical vector.
}
