% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{as_hedgehog}
\alias{as_hedgehog}
\title{Convert a quickcheck generator to a hedgehog generator}
\usage{
as_hedgehog(generator)
}
\arguments{
\item{generator}{A \code{quickcheck_generator} generator}
}
\value{
A \code{quickcheck_generator} generator
}
\description{
Convert a quickcheck generator to a hedgehog generator
}
\examples{
is_even <-
  \(a) a \%\% 2L == 0L

gen_powers_of_two <-
  as_hedgehog(integer_()) |> hedgehog::gen.with(\(a) 2 ^ a)

for_all(
  a = gen_powers_of_two,
  property = \(a) is_even(a) |> testthat::expect_true()
)
}
