% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{break_condition_message}
\alias{break_condition_message}
\title{Retrieve the condition message and breaks it in a character vector based on newline character.
This is helpful since quickalerts have been defined using only the message argument.
Therefore the body of the message is condensed with the message (we work with one single message).}
\usage{
break_condition_message(condition)
}
\arguments{
\item{condition}{condition object.}
}
\description{
Retrieve the condition message and breaks it in a character vector based on newline character.
This is helpful since quickalerts have been defined using only the message argument.
Therefore the body of the message is condensed with the message (we work with one single message).
}
