% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cancer_list}
\alias{cancer_list}
\title{cancer_list <- df transformed in list with added a list inside of it.}
\format{
list version of df with a list nested into it.
}
\usage{
cancer_list
}
\description{
cancer_list <- df transformed in list with added a list inside of it.
}
\keyword{datasets}
