% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_arg}
\alias{check_arg}
\title{Check function used to check the single arg.}
\usage{
check_arg(arg, exp_type, null, flag, with, param)
}
\arguments{
\item{arg}{Argument to check.}

\item{exp_type}{String reporting the type expectation on that arg.}

\item{null}{Logical vector of the same length of args or of length 1, indicating whether the correspondent args can be null.
If of length one the vector is recycled to args length.}

\item{flag}{Logical vector of the same length of args or of length 1, indicating whether the correspondent args have to be of length 1.
If of length one the vector is recycled to args length}

\item{with}{Character vector indicating which functions to use to get the actual types/class on which the expectation must be checked.
Must be of the same length of args or of length 1. If of length one the vector is recycled to args length.
Possible values are:
'typeof': declare the use of the "typeof" function,
'class': declare the use of the "class" function,
'integerish': declare a check internally performed via class_integerish returning "integerish" for rlang-defined-integerish objects.}

\item{param}{Name of the parameter.}
}
\description{
Check function used to check the single arg.
}
