% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R, R/make_test_function.R
\name{check_columns_copresence}
\alias{check_columns_copresence}
\alias{test_columns_copresence}
\title{Check duplicated column names in two dataframes}
\usage{
check_columns_copresence(
  df1,
  df2,
  columns = NULL,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)

test_columns_copresence(
  df1,
  df2,
  columns = NULL,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{columns}{Character vector of columns to consider. If NULL, the default, all columns are considered. No check is done for not existent columns.}

\item{df1_arg}{String specifying how to address first df in the raised messages (default "df1").}

\item{df2_arg}{String specifying how to address second df in the raised messages (default "df2").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{String to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Check duplicated column names in two dataframes
}
\details{
The function does not check for the existence of the passed columns value.
Therefore is possible to pass names not found in both dataframes and no error is raised.
}
