% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R, R/make_test_function.R
\name{check_columns_presence}
\alias{check_columns_presence}
\alias{test_columns_presence}
\title{Check the presence of one or multiple columns in a dataframe}
\usage{
check_columns_presence(
  df,
  columns,
  df_arg = "df",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)

test_columns_presence(
  df,
  columns,
  df_arg = "df",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{Character vector reporting the column/s name.}

\item{df_arg}{String specifying how to address df in the alerts (default "df").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{String added at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Check the presence of one or multiple columns in a dataframe
}
