% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_copresence_dfs}
\alias{check_copresence_dfs}
\title{Perform an ordered cross-checking between the values of two columns of two dataframes.}
\usage{
check_copresence_dfs(
  df1,
  df2,
  col,
  direction = "first_in_second",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{col}{character vector referring the name of the columns to consider. If of length one
the single name is assumed for both dataframes. If of length two the first name is assumed
for the first dataframe and the other for the second.}

\item{direction}{string equal to one of 'first_in_second', 'second_in_first' or 'bidirectional'.
Set the direction of the comparison.}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be ignored
most of times because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired). So to point to the calling frame of this function you have to set 4.}
}
\description{
The function allows to check the presence of all values of the selected column of one dataframe
in the other. If direction equal 'bidirectional' a bidirectional check is performed.
}
