% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R, R/make_test_function.R
\name{check_dataframe}
\alias{check_dataframe}
\alias{test_dataframe}
\title{Ensemble check for a dataframe}
\usage{
check_dataframe(
  df,
  df_arg = "df",
  null_check = FALSE,
  dim_check = FALSE,
  n_exact_cols = NULL,
  n_min_cols = NULL,
  n_max_cols = NULL,
  predicate = NULL,
  inverse = FALSE,
  columns = NULL,
  cols_key = FALSE,
  rm_na_key = FALSE,
  cols_na = TRUE
)

test_dataframe(
  df,
  df_arg = "df",
  null_check = FALSE,
  dim_check = FALSE,
  n_exact_cols = NULL,
  n_min_cols = NULL,
  n_max_cols = NULL,
  predicate = NULL,
  inverse = FALSE,
  columns = NULL,
  cols_key = FALSE,
  rm_na_key = FALSE,
  cols_na = TRUE
)
}
\arguments{
\item{df}{Dataframe to check.}

\item{df_arg}{String reporting how to adress the dataframe in the alerts (default "df").}

\item{null_check}{Boolean, indicating whether to check for a NULL value (default FALSE).}

\item{dim_check}{Boolean, indicating whether to check for 0 dimensional dataframe (both columns and rows) (default FALSE).}

\item{n_exact_cols}{Integer(ish), indicating the exact expected number of columns (default NULL).}

\item{n_min_cols}{Integer(ish), indicating the minimum expected number of columns (default NULL).}

\item{n_max_cols}{Integer(ish), indicating the maximum expected number of columns (default NULL).}

\item{predicate}{Predicate function that all df columns have to satisfy (default NULL).}

\item{inverse}{Boolean, whether the predicate must be NOT satisfied for all the elements for a successful check (default FALSE).}

\item{columns}{Character vector reporting the columns names on which the presence check and the additional optional specific check are perfomed.}

\item{cols_key}{Boolean, should be the columns specified in "columns" made of only unique values (default FALSE).}

\item{rm_na_key}{Boolean, should the NAs not considered during the "key" check (default FALSE).}

\item{cols_na}{Boolean, specifying whether the columns specified in "columns" can have NAs (default TRUE).}
}
\description{
Allow to test for different proprieties of a dataframe.
The properties are tested in the order they compare as parameters function.
}
