% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_integerish}
\alias{check_integerish}
\title{Helper to convert the output of is_integerish to strings.
In detail if x is integerish, the function return "integerish", otherwise "non_integerish".}
\usage{
check_integerish(x)
}
\arguments{
\item{x}{object on wich is_integerish is called.}
}
\description{
Helper to convert the output of is_integerish to strings.
In detail if x is integerish, the function return "integerish", otherwise "non_integerish".
}
