% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_key}
\alias{check_key}
\title{Checks if the specified column of the dataframe is suitable as key.}
\usage{
check_key(df, key)
}
\arguments{
\item{df}{dataframe.}

\item{key}{character reporting the column name.}
}
\description{
Checks if the specified column of the dataframe is suitable as key.
}
