% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{check_len_args}
\alias{check_len_args}
\title{Internal check for .*_len arguments of check_length_vec and check_length_list}
\usage{
check_len_args(exact_len, min_len, max_len)
}
\arguments{
\item{exact_len}{Integer indicating the exact expected vector length (default NULL).}

\item{min_len}{Integer indicating the minimum expected vector length (default NULL).}

\item{max_len}{Integer indicating the maximum expected vector length (default NULL).}
}
\description{
Internal check for .*_len arguments of check_length_vec and check_length_list
}
