% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_list.R, R/make_test_function.R
\name{check_length_list}
\alias{check_length_list}
\alias{test_length_list}
\title{Check the length of a list}
\usage{
check_length_list(
  x,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  flatten = FALSE,
  xarg = "list",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)

test_length_list(
  x,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  flatten = FALSE,
  xarg = "list",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{x}{List to check.}

\item{exact_len}{Integer indicating the exact expected vector length (default NULL).}

\item{min_len}{Integer indicating the minimum expected vector length (default NULL).}

\item{max_len}{Integer indicating the maximum expected vector length (default NULL).}

\item{flatten}{Boolean, whether to flat out the list before doing the check (default TRUE).}

\item{xarg}{String, indicating how to refer x in the alert messages (default "list").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Checks whether the expectations on the list length hold.
}
