% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_vecs.R, R/make_test_function.R
\name{check_matching_vecs}
\alias{check_matching_vecs}
\alias{test_matching_vecs}
\title{Check the unordered equality of two vectors}
\usage{
check_matching_vecs(
  vec1,
  vec2,
  na_rm = FALSE,
  coerce = FALSE,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n_evaluation_frame = 0,
  ...
)

test_matching_vecs(
  vec1,
  vec2,
  na_rm = FALSE,
  coerce = FALSE,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n_evaluation_frame = 0,
  ...
)
}
\arguments{
\item{vec1}{First vector.}

\item{vec2}{Second vector.}

\item{na_rm}{Boolean indicating if NA must be excluded prior checking (default FALSE).}

\item{coerce}{Boolean indicating whether the two vectors can be coerced during the check (default FALSE).
If FALSE a check on vector classes is performed before the "main" check. This check is performed using the identical function.
Therefore the vector classes must be perfectly identical in order to pass the check.}

\item{vec1_arg}{String indicating how to address vec1 in the raised message (default 'vec1').}

\item{vec2_arg}{String indicating how to address vec2 in the raised message (default 'vec2)'.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
The equality is checked thought the matching \%in\% operator in both senses.
The "\%in\%" operator uses automatic coercion. To control this behavior see the coerce parameter.
The vectors orders as well as the presence of repeated values have no importance.
If the order and repetitions have a role see "check_equality_vecs" function.
}
