% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R, R/make_test_function.R
\name{check_names}
\alias{check_names}
\alias{test_names}
\title{Perform various checks on object names}
\usage{
check_names(
  x,
  xarg = "x",
  null_check = FALSE,
  zero_len_check = FALSE,
  na_check = FALSE,
  empty_string_check = FALSE,
  unique = FALSE,
  na_rm_unique = FALSE,
  include = NULL,
  exclude = NULL,
  what = "names"
)

test_names(
  x,
  xarg = "x",
  null_check = FALSE,
  zero_len_check = FALSE,
  na_check = FALSE,
  empty_string_check = FALSE,
  unique = FALSE,
  na_rm_unique = FALSE,
  include = NULL,
  exclude = NULL,
  what = "names"
)
}
\arguments{
\item{x}{Object with one of the attributes accepted by the param "what".}

\item{xarg}{String reporting how to adress the object in the alerts (default "x").}

\item{null_check}{Boolean, indicating whether to check for a NULL vec (default FALSE).}

\item{zero_len_check}{Boolean, indicating whether to check for a zero-length vec (default FALSE).}

\item{na_check}{Boolean, indicating whether to check for absence of NAs (default FALSE).}

\item{empty_string_check}{Boolean, indicating whether to check for absence of empty strings "" (default FALSE).}

\item{unique}{Boolean, should be vec a vector of unique values (default FALSE).}

\item{na_rm_unique}{Boolean, should the NAs be removed for the unique-check (default FALSE).}

\item{include}{Vector of values wich presence is checked (default NULL).}

\item{exclude}{Vector of values wich absence is checked (default NULL).}

\item{what}{Type of name attribute to check, one of "names", "colnames" or "rownames" (default "names").}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Perform various checks on object names
}
