% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_nrow}
\alias{check_nrow}
\title{Checks equal number of rows between two dataframes}
\usage{
check_nrow(
  df1,
  df2,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{df1_arg}{string specifying how to address df1 in the raised messages (default "df1").}

\item{df2_arg}{string specifying how to address df2 in the raised messages (default "df2").}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be most of
times ignored because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired). So to point to the calling frame of this function you have to set 4.}
}
\description{
Checks equal number of rows between two dataframes
}
