% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R, R/make_test_function.R
\name{check_nrow_dfs}
\alias{check_nrow_dfs}
\alias{test_nrow_dfs}
\title{Check equal number of rows between two dataframes}
\usage{
check_nrow_dfs(
  df1,
  df2,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)

test_nrow_dfs(
  df1,
  df2,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{df1_arg}{String specifying how to address first df in the raised messages (default "df1").}

\item{df2_arg}{String specifying how to address second df in the raised messages (default "df2").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Check equal number of rows between two dataframes
}
