% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_predicate_columns}
\alias{check_predicate_columns}
\title{check if all predicate result on vector columns are TRUE}
\usage{
check_predicate_columns(
  df,
  predicate,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\description{
check if all predicate result on vector columns are TRUE
}
