% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_presence_values}
\alias{check_presence_values}
\title{Checks the presence of the specified values in a vector.}
\usage{
check_presence_values(
  vec,
  values,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{vec}{Vector to check.}

\item{values}{character vector of values searched in vec.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the function frame. So it's possible to simply points
to upper frames (as well as to below frames but is not recommended).}
}
\description{
Checks the presence of the specified values in a vector.
}
