% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_vectors.R
\name{check_unordered_equality_vecs}
\alias{check_unordered_equality_vecs}
\title{Checks the unordered equality of two vectors.}
\usage{
check_unordered_equality_vecs(
  vec1,
  vec2,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n.evaluation_frame = 0,
  ...
)
}
\arguments{
\item{vec1}{First vector.}

\item{vec2}{Second vector.}

\item{vec1_arg}{string indicating how to address vec1 in the raised message (default 'vec1').}

\item{vec2_arg}{string indicating how to address vec2 in the raised message (default 'vec2)'.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{n.evaluation_frame}{numeric, defines the number of stack frame to look up for the evaluation of the glue expressions of the alert message.
The default value points to the frame of this function.}

\item{...}{not of direct use.}
}
\value{
invisible NULL
}
\description{
The equality is checked thought the \%in\% operator in both senses.
The vectors orders as well as the presence of repeated values have no importance.
If the order and the repetitions have a role see "check_equality_vecs" function.
}
