% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vec.R, R/make_test_function.R
\name{check_vector}
\alias{check_vector}
\alias{test_vector}
\title{Ensemble check for an atomic vector}
\usage{
check_vector(
  vec,
  vec_arg = "vec",
  null_check = FALSE,
  zero_len_check = FALSE,
  na_check = FALSE,
  empty_string_check = FALSE,
  predicate = NULL,
  inverse = FALSE,
  sorted = FALSE,
  decreasing = FALSE,
  unique = FALSE,
  na_rm_unique = FALSE,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  unique_len = FALSE,
  na_rm_len = FALSE,
  include = NULL,
  exclude = NULL
)

test_vector(
  vec,
  vec_arg = "vec",
  null_check = FALSE,
  zero_len_check = FALSE,
  na_check = FALSE,
  empty_string_check = FALSE,
  predicate = NULL,
  inverse = FALSE,
  sorted = FALSE,
  decreasing = FALSE,
  unique = FALSE,
  na_rm_unique = FALSE,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  unique_len = FALSE,
  na_rm_len = FALSE,
  include = NULL,
  exclude = NULL
)
}
\arguments{
\item{vec}{Vector to check. Must be an atomic vector (see function description).}

\item{vec_arg}{String reporting how to adress the vec in the alerts (default "vec").}

\item{null_check}{Boolean, indicating whether to check for a NULL vec (default FALSE).}

\item{zero_len_check}{Boolean, indicating whether to check for a zero-length vec (default FALSE).}

\item{na_check}{Boolean, indicating whether to check for absence of NAs (default FALSE).}

\item{empty_string_check}{Boolean, indicating whether to check for absence of empty strings "" (default FALSE).}

\item{predicate}{A predicate function to apply to all the elements of vec (default NULL).}

\item{inverse}{Boolean, whether the predicate must be NOT satisfied for all the elements for a successful check (default FALSE).}

\item{sorted}{Boolean, is vec sorted? (default FALSE).}

\item{decreasing}{Boolean, should the sort check be performed in decreasing order? (default FALSE).}

\item{unique}{Boolean, should be vec a vector of unique values (default FALSE).}

\item{na_rm_unique}{Boolean, should the NAs be removed for the unique-check (default FALSE).}

\item{exact_len}{Integer(ish), indicating the exact expected vector length (default NULL).}

\item{min_len}{Integer(ish), indicating the minimum expected vector length (default NULL).}

\item{max_len}{Integer(ish), indicating the maximum expected vector length (default NULL).}

\item{unique_len}{Boolean, should the length-check be performed on only the unique values of vec (default FALSE).}

\item{na_rm_len}{Boolean, should the NAs be removed for the length-check (default FALSE).}

\item{include}{Vector of values wich presence is checked (default NULL).}

\item{exclude}{Vector of values wich absence is checked (default NULL).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Allow to test for different proprieties of an atomic vector. For the definition of “atomic”, see \link[base]{is.atomic}.
The properties are tested in the order they compare as parameters function.
}
