% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clean_vec}
\alias{clean_vec}
\title{Manage sorting, NAs and duplicates removal for a vector.}
\usage{
clean_vec(vec, na_rm = FALSE, unique = FALSE, sort = FALSE, decreasing = FALSE)
}
\arguments{
\item{vec}{Input vector.}

\item{na_rm}{Boolean indicating if NA must be removed.}

\item{unique}{Boolean indicating whether to perform the check only on unique values (default FALSE).}

\item{sort}{Boolean, whether to sort the vectors before performing the check (default FALSE).}

\item{decreasing}{Boolean indicating the sorting order.}
}
\value{
The cleaned vector.
}
\description{
Manage sorting, NAs and duplicates removal for a vector.
}
