% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{control_recycle}
\alias{control_recycle}
\title{Helper of check_primitive_types to control the recycle of expected types and classes.
Vector2 is recycled according to the numeric correspondence vector.}
\usage{
control_recycle(vector1, vector2, numeric_correspondence)
}
\arguments{
\item{vector1}{first vector.}

\item{vector2}{second vector.}

\item{numeric_correspondence}{number of times the elements of vector2 must be repeated in order.}
}
\description{
Helper of check_primitive_types to control the recycle of expected types and classes.
Vector2 is recycled according to the numeric correspondence vector.
}
