% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasversal_internals.R
\name{core_length_test}
\alias{core_length_test}
\title{Internal function that does the checking for the check_length_.* funcs and builds the alert_message}
\usage{
core_length_test(len_object, exact_len, min_len, max_len, string_object)
}
\arguments{
\item{len_object}{Length of the object to check and report in the alert messages.}

\item{exact_len}{Integer indicating the exact expected vector length (default NULL).}

\item{min_len}{Integer indicating the minimum expected vector length (default NULL).}

\item{max_len}{Integer indicating the maximum expected vector length (default NULL).}

\item{string_object}{String to adress the object in the alert messages.}
}
\value{
The alert message or NULL if no alert must be raised.
}
\description{
Internal function that does the checking for the check_length_.* funcs and builds the alert_message
}
