% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_list.R
\name{custom_identical}
\alias{custom_identical}
\title{Helper of check_uniform_list.}
\usage{
custom_identical(x, y)
}
\arguments{
\item{x}{first element.}

\item{y}{second element.}
}
\description{
Helper of check_uniform_list.
}
