% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{data_frame_}
\alias{data_frame_}
\title{Data frame generator}
\usage{
data_frame_(..., rows = c(1L, 10L))
}
\arguments{
\item{...}{A set of name-value pairs with the values
being vector generators.}

\item{rows}{Number of rows of the generated data frame.
If \code{rows} is a single number all data frames will have
this number of rows. If \code{rows} is a numeric vector of
length 2 it will produce data frames with rows between
a minimum and maximum, inclusive. For example
\code{rows = c(1L, 10L)} would produce data frames with rows
between 1 and 10.}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
Construct data frame generators in a similar way to
\code{base::data.frame}.
}
\examples{
data_frame_(a = integer_()) |> show_example()
data_frame_(a = integer_(), b = character_(), rows = 5L) |> show_example()
}
