% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{default_alert_messages}
\alias{default_alert_messages}
\title{default_alert_messages --> list with the default messages raised by the checking functions.}
\format{
list of character vector or rarely list (for function that raise different version of checks)
}
\usage{
default_alert_messages
}
\description{
default_alert_messages --> list with the default messages raised by the checking functions.
}
\keyword{datasets}
