% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{detect_type_condition}
\alias{detect_type_condition}
\title{Inspect and return the "classic" nature of conditions. Helper of impose_logical_behavior for re-signalling conditions.
Here classic means error, warning or message classes.}
\usage{
detect_type_condition(condition)
}
\arguments{
\item{condition}{condition object.}
}
\description{
Inspect and return the "classic" nature of conditions. Helper of impose_logical_behavior for re-signalling conditions.
Here classic means error, warning or message classes.
}
