% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{df_row}
\alias{df_row}
\title{df_row --> rows have changed}
\format{
has half rows but the same columns of df
}
\usage{
df_row
}
\description{
df_row --> rows have changed
}
\keyword{datasets}
