% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{equal_length}
\alias{equal_length}
\title{Modify length of multiple generators}
\usage{
equal_length(..., len = 1L)
}
\arguments{
\item{...}{Generators}

\item{len}{Length of the generated vectors. If \code{len} is a
single number all vectors will have this length. If
\code{len} is a numeric vector of length 2 it will produce
vectors with lengths between a minimum and maximum,
inclusive. For example \code{len = c(1L, 10L)} would produce
vectors with lengths between 1 and 10.}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
This will set the length of multiple generators to be the
same.
}
