% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_all.R
\name{for_all}
\alias{for_all}
\title{For all}
\usage{
for_all(
  ...,
  property,
  tests = get_tests(),
  size = get_size(),
  shrinks = get_shrinks(),
  discards = get_discards()
)
}
\arguments{
\item{...}{Dots}

\item{property}{Property}

\item{tests}{Tests}

\item{size}{Size}

\item{shrinks}{Shrinks}

\item{discards}{Discards}
}
\description{
For all
}
