% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{generate_alertfunc_list}
\alias{generate_alertfunc_list}
\title{Generates a list of aliases of main cli alert functions with intuitive names.}
\usage{
generate_alertfunc_list()
}
\description{
Generates a list of aliases of main cli alert functions with intuitive names.
}
