% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{generate_cli_alert_list}
\alias{generate_cli_alert_list}
\title{Generates a list of main cli alert functions with informative names.}
\usage{
generate_cli_alert_list()
}
\description{
Generates a list of main cli alert functions with informative names.
}
