% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{generate_cli_sign_list}
\alias{generate_cli_sign_list}
\title{Generates a list of cli bullet signs with informative names.}
\usage{
generate_cli_sign_list()
}
\description{
Generates a list of cli bullet signs with informative names.
}
