% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{generate_header}
\alias{generate_header}
\title{Function to construct a default header message if not passed via dots.}
\usage{
generate_header(header, default_header)
}
\arguments{
\item{header}{takes in the header argument of the checking function.}

\item{default_header}{string that is used as default message.}
}
\description{
Function to construct a default header message if not passed via dots.
}
