% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generator.R
\name{generate_message}
\alias{generate_message}
\title{Construct a default alert message if not provided (if NULL).}
\usage{
generate_message(alert_message, default_message)
}
\arguments{
\item{alert_message}{takes in the alert_message argument of the checking function.}

\item{default_message}{character vector that is used as default message. Follow cli intax rules.}
}
\description{
Construct a default alert message if not provided (if NULL).
}
