% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_check.R
\name{generic_check}
\alias{generic_check}
\title{Generic quickcheck constructor.}
\usage{
generic_check(
  expr,
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{expr}{Expression to test. Must return a single boolean value.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Invisible NULL
}
\description{
Allows to test generic hypothesis and raising conditions while having all the additional features of quickcheck.
The condition is raised if the supplied expression evaluates to TRUE.
}
