% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_errors_predicate}
\alias{get_errors_predicate}
\title{Retrieve and returns the elements in error.
Wants in input a boolena indicating if the vector/list has empty names.
In this last case it returns a value 1 as error.}
\usage{
get_errors_predicate(object, log_empty_names, predicate_log)
}
\arguments{
\item{object}{Vector or list.}

\item{log_empty_names}{Boolean indicating if the object on which the predicate is checked has empty names.}

\item{predicate_log}{Boolean vector with the result of the application of the predicate.}
}
\description{
Retrieve and returns the elements in error.
Wants in input a boolena indicating if the vector/list has empty names.
In this last case it returns a value 1 as error.
}
