% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{get_name}
\alias{get_name}
\title{Helper of impose_loop_beahvior to check and get the name of the element of a list.
If not set in broad sense it get converted to nameroot pasted to iteration.}
\usage{
get_name(name, nameroot, iteration)
}
\arguments{
\item{name}{name to check}

\item{nameroot}{basename of the unnamed element.}

\item{iteration}{number added to the nameroot to constitute the default element name.}
}
\description{
Helper of impose_loop_beahvior to check and get the name of the element of a list.
If not set in broad sense it get converted to nameroot pasted to iteration.
}
