\name{as.assertion}
\alias{as.assertion}
\alias{fun}
\title{Helper functions for quikcheck}
\description{\code{as.assertion} allows to convert any testthat expectation into a assertion appropriate for use with the function \code{unit.test}, \code{fun} id equivalent to Curry but add a little syntactic sugar}
\usage{
as.assertion(expectation)
fun(a.call)}
\arguments{
\item{expectation}{}
\item{a.call}{}
}