% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_accumulation_behavior}
\alias{impose_accumulation_behavior}
\title{Imposes an accumulation behavior of the checking function in a loop scenario, in which
the checking function can raise alert that are stored in a list and finally at the end of
the loop reported. Set to works exclusively with message type in order to distinguish
quickalert alert from normal alert.}
\usage{
impose_accumulation_behavior(
  expr,
  type = "error",
  header = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{expr}{check function call.}

\item{type}{type of the accumulated final alert if any.}

\item{header}{string to add as the head of the accumulated alert list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run. The default points
to this function frame.}
}
\description{
Imposes an accumulation behavior of the checking function in a loop scenario, in which
the checking function can raise alert that are stored in a list and finally at the end of
the loop reported. Set to works exclusively with message type in order to distinguish
quickalert alert from normal alert.
}
