% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_additional_alert}
\alias{impose_additional_alert}
\title{Adds alert messages at the start or end of raised quickalert messages}
\usage{
impose_additional_alert(
  expr,
  message,
  margin = 1,
  raise = "error",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{expr}{Check function call. It's important that the checking function raises warnings or messages and not errors.}

\item{message}{String to add to the alert message.}

\item{margin}{Numeric equal to 1 or 2, indicating where to add the additional message (start or end respectively).}

\item{raise}{Type of the accumulated final alert if any.}

\item{n_evaluation_frame}{Intergerish, defines the number of stack frame to look down for the evaluation
of the glue expressions of the alert message. The default (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert is of class "quickalert" (default TRUE).}

\item{...}{To pass additional parameters to alert_generator function.}
}
\value{
invisible NULL
}
\description{
This wrapper works only with messages and warnings and not errors.
Note that this is not enforced by the function and therefore the responsability fall on the user.
}
