% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_single_vec.R
\name{internal_check_empty_vec}
\alias{internal_check_empty_vec}
\title{Internal check of check_empty_vec}
\usage{
internal_check_empty_vec(
  vec,
  null = TRUE,
  len = TRUE,
  na = TRUE,
  empty_string = TRUE,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{null}{Boolean, indicating whether to perform the check for NULL value (default TRUE).}

\item{len}{Boolean, indicating whether to perform the check for 0 length (default TRUE).}

\item{na}{Boolean, indicating whether to perform the check for NAs (default TRUE).}

\item{empty_string}{Boolean, indicating whether to perform the check for empty string "" (default TRUE).}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Internal check of check_empty_vec
}
