% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_cross_vecs.R
\name{internal_check_identical_vecs}
\alias{internal_check_identical_vecs}
\title{Internal check of check_identical_vecs}
\usage{
internal_check_identical_vecs(
  vec1,
  vec2,
  na_rm = FALSE,
  unique = FALSE,
  sort = FALSE,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n_evaluation_frame = 0,
  ...
)
}
\arguments{
\item{vec1}{First vector.}

\item{vec2}{Second vector.}

\item{na_rm}{Boolean indicating if NA must be excluded prior checking (default FALSE).}

\item{unique}{Boolean indicating whether to perform the check only on unique values (default FALSE).}

\item{sort}{Boolean, whether to sort the vectors before performing the check (default FALSE).}

\item{vec1_arg}{String indicating how to address vec1 in the raised message (default 'vec1').}

\item{vec2_arg}{String indicating how to address vec2 in the raised message (default 'vec2)'.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Internal check of check_identical_vecs
}
