% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_single_list.R
\name{internal_check_list}
\alias{internal_check_list}
\title{Internal version of check_list}
\usage{
internal_check_list(
  x,
  xarg = "list",
  flatten = FALSE,
  null_check = FALSE,
  zero_len_check = FALSE,
  uniform = FALSE,
  predicate = NULL,
  inverse = FALSE,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL
)
}
\arguments{
\item{x}{List to check.}

\item{xarg}{String, indicating how to refer x in the alert messages (default "list").}

\item{flatten}{Single boolean, whether to flat out the list before doing the check (default TRUE).}

\item{null_check}{Boolean, indicating whether to check for a NULL value (default FALSE).}

\item{zero_len_check}{Boolean, indicating whether to check for a zero-length list (default FALSE).}

\item{uniform}{Boolean, indicating whether to check for identical classed list elements (default FALSE).}

\item{predicate}{A predicate function, usually an "is.something" function (default NULL).}

\item{inverse}{Boolean, whether the predicate must not be satisfied for all the elements (default FALSE).}

\item{exact_len}{Integer(ish), indicating the exact expected list length (default NULL).}

\item{min_len}{Integer(ish), indicating the minimum expected list length (default NULL).}

\item{max_len}{Integer(ish), indicating the maximum expected list length (default NULL).}
}
\description{
Internal version of check_list
}
