% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_single_list.R
\name{internal_check_predicate_list}
\alias{internal_check_predicate_list}
\title{Internal version of check_predicate_list}
\usage{
internal_check_predicate_list(
  x,
  predicate,
  inverse = FALSE,
  flatten = FALSE,
  xarg = "list",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{x}{List to check.}

\item{predicate}{A predicate function, usually an "is.something" function.}

\item{inverse}{Boolean, whether to invert the check direction in the sense that the predicate must be not satisfied for all the elements (default FALSE).}

\item{flatten}{Boolean, whether to flat out the list before doing the check (default TRUE).}

\item{xarg}{String, indicating how to refer x in the alert messages (default "list").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Internal version of check_predicate_list
}
