% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{is_empty_vec}
\alias{is_empty_vec}
\title{Helper of check_empty_vec in which all "empty possibilities" are tested for each value.
Return a single boolean (FALSE even if one value is empty, TRUE otherwise).}
\usage{
is_empty_vec(vec)
}
\arguments{
\item{vec}{vector to check}
}
\description{
Helper of check_empty_vec in which all "empty possibilities" are tested for each value.
Return a single boolean (FALSE even if one value is empty, TRUE otherwise).
}
