% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_}
\alias{list_}
\title{List generator}
\usage{
list_(...)
}
\arguments{
\item{...}{Quickcheck generators}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
This will generate lists with contents corresponding to the values generated
by the input generators.
}
\examples{
list_(integer_(), logical_()) |> show_example()
list_(a = any_vector(), b = any_vector()) |> show_example()
}
