% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_of}
\alias{list_of}
\title{Variable length list generator}
\usage{
list_of(generator, len = 1L)
}
\arguments{
\item{generator}{Quickcheck generator}

\item{len}{Length of the generated vectors. If \code{len} is a
single number all vectors will have this length. If
\code{len} is a numeric vector of length 2 it will produce
vectors with lengths between a minimum and maximum,
inclusive. For example \code{len = c(1L, 10L)} would produce
vectors with lengths between 1 and 10.}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
This will generate lists with all values coming from a single input generator.
}
\examples{
list_of(integer_(), len = 10L) |> show_example()
}
