% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{logical_}
\alias{logical_}
\title{Logical generator}
\usage{
logical_(len = c(1L, 10L), any_na = FALSE)
}
\arguments{
\item{len}{Length of the generated vectors. If \code{len} is a
single number all vectors will have this length. If
\code{len} is a numeric vector of length 2 it will produce
vectors with lengths between a minimum and maximum,
inclusive. For example \code{len = c(1L, 10L)} would produce
vectors with lengths between 1 and 10. To produce empty vectors
set \code{len = 0L} or a range like \code{len = c(0L, 10L)}.}

\item{any_na}{Whether \code{NA} values should be allowed.}
}
\value{
A \code{quickcheck_generator} object.
}
\description{
A generator for logical vectors.
}
\examples{
logical_() |> show_example()
logical_(len = 10L, any_na = TRUE) |> show_example()
}
