% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_function.R
\name{make_test_function}
\alias{make_test_function}
\title{Make test function from check_function}
\usage{
make_test_function(check_fun, env = parent.frame())
}
\arguments{
\item{check_fun}{String reporting the name of the check function to convert.}

\item{env}{Environment set in the function object (for scoping purposes).
Default the caller env that is the package NAMESPACE at building time.}
}
\value{
The test function.
}
\description{
Make test function from check_function
}
