% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numeric_}
\alias{numeric_}
\alias{numeric_bounded}
\alias{numeric_left_bounded}
\alias{numeric_right_bounded}
\alias{numeric_positive}
\alias{numeric_negative}
\title{Numeric generators}
\usage{
numeric_(len = 1L, frac_na = 0)

numeric_bounded(left, right, len = 1L, frac_na = 0)

numeric_left_bounded(left, len = 1L, frac_na = 0)

numeric_right_bounded(right, len = 1L, frac_na = 0)

numeric_positive(len = 1L, frac_na = 0)

numeric_negative(len = 1L, frac_na = 0)
}
\arguments{
\item{len}{Length of the generated vectors. If \code{len} is a
single number all vectors will have this length. If
\code{len} is a numeric vector of length 2 it will produce
vectors with lengths between a minimum and maximum,
inclusive. For example \code{len = c(1L, 10L)} would produce
vectors with lengths between 1 and 10.}

\item{frac_na}{The approximate fraction of \code{NA} values to
be included in the vector. A fraction of 0, the default,
will include no \code{NA} values. A fraction of 0.5 would
include around 50\% \code{NA} values.}

\item{left}{The minimum possible value for generated
numbers, inclusive.}

\item{right}{The maximum possible value for generated
numbers, inclusive.}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
A set of generators for numeric vectors. Numeric vectors
can be either integer or double vectors.
}
