% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_of.R
\name{one_of}
\alias{one_of}
\title{Randomly choose between generators}
\usage{
one_of(..., prob = NULL)
}
\arguments{
\item{...}{A set of unnamed generators.}

\item{prob}{A vector of probability weights for obtaining the elements of the
vector being sampled.}
}
\value{
A \code{quickcheck_generator} object.
}
\description{
Randomly choose between generators
}
\examples{
one_of(integer_(), character_()) \%>\% show_example()
one_of(constant(NULL), logical_(), prob = c(0.1, 0.9)) \%>\% show_example()
}
