% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{qadf1}
\alias{qadf1}
\title{Dataframe for check on double dataframes}
\format{
Subset of df with only visit_number, sex and age columns.
}
\usage{
qadf1
}
\description{
Dataframe for check on double dataframes
}
\keyword{datasets}
