% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{qadf2}
\alias{qadf2}
\title{Dataframe for check on double dataframes}
\format{
Subset of df with only visit_number, bmi and height columns.
}
\usage{
qadf2
}
\description{
Dataframe for check on double dataframes
}
\keyword{datasets}
