% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcheck-package.R
\docType{package}
\name{quickcheck-package}
\alias{quickcheck}
\alias{quickcheck-package}
\title{quickcheck: Implementation of common cheks for typical data structure}
\description{
Implements checks that are performed on a single dataframe or a couple of dataframes, which can be useful to enhance the robustness of other functions.
}
\author{
\strong{Maintainer}: Michele Avagliano \email{avaglianomichele@outlook.it}

}
\keyword{internal}
