% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generator.R
\name{raise_eval_frame}
\alias{raise_eval_frame}
\title{Raises the evaluation frame number by x if not 0. It's purpose is to allow to skip stack frames.}
\usage{
raise_eval_frame(n.evaluation_frame, up)
}
\arguments{
\item{n.evaluation_frame}{Integer indicating the stack frame in which the alert glue expressions are evaluated in respect to where is generated.
So it indicates the number of frames to look down in the stack. By default is zero plus two.
The addition is used in order to start counting from zero from the frame of the function that uses this function.}

\item{up}{The number added to the frame number.}
}
\description{
Raises the evaluation frame number by x if not 0. It's purpose is to allow to skip stack frames.
}
