% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasversal_internals.R
\name{raise_length_alert}
\alias{raise_length_alert}
\title{Helper to raise alerts of check_length_* funcs.}
\usage{
raise_length_alert(
  raise,
  combined_message,
  alert_message,
  n_evaluation_frame,
  quickalert,
  ...
)
}
\arguments{
\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{combined_message}{alert build by core_length_test function.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Integerish, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{Logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Helper to raise alerts of check_length_* funcs.
}
