% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{rec_flatten_list}
\alias{rec_flatten_list}
\title{Flatten recursively a list}
\usage{
rec_flatten_list(l, num = NULL, till_flat = TRUE, nam_spec = "{outer}_{inner}")
}
\arguments{
\item{l}{list to flatten.}

\item{num}{number of layers to remove.}

\item{till_flat}{logical; flat out completely a list (default = TRUE).}

\item{nam_spec}{If both inner and outer names are present, control how they are combined.
It has to be a glue specification that uses variables inner and outer. Default = {outer}_{inner}.}
}
\value{
Returns the flattened list.
}
\description{
The function allows to remove a desired number of layers or to flat out a list independently of its structure.
}
