% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{reorder_ids_metadata}
\alias{reorder_ids_metadata}
\title{The function initially checks for the presence of sampleid_col in both dfs raising
error (if missing in data) and warning (if missing in metadata), and then proceed to
reorder metadata rows if necessary according to the data ids.}
\usage{
reorder_ids_metadata(
  data,
  metadata,
  sampleid_col,
  data_arg = "data",
  metadata_arg = "metadata"
)
}
\arguments{
\item{data}{dataframe that set the order with a column that can act as key.}

\item{metadata}{dataframe that contain the same key and which order is corrected if necessary.}

\item{sampleid_col}{character reporting the key column name.}

\item{data_arg}{string used for data argument in alerts.}

\item{metadata_arg}{string used for metadata argument in alerts.}
}
\value{
The reordered metadata
}
\description{
The function initially checks for the presence of sampleid_col in both dfs raising
error (if missing in data) and warning (if missing in metadata), and then proceed to
reorder metadata rows if necessary according to the data ids.
}
