% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{select_check_func}
\alias{select_check_func}
\title{Select the correct check func based on with value.}
\usage{
select_check_func(with)
}
\arguments{
\item{with}{Character vector indicating which functions to use to get the actual types/class on which the expectation must be checked.
Must be of the same length of args or of length 1. If of length one the vector is recycled to args length.
Possible values are:
'typeof': declare the use of the "typeof" function,
'class': declare the use of the "class" function,
'integerish': declare a check internally performed via class_integerish returning "integerish" for rlang-defined-integerish objects.}
}
\description{
Select the correct check func based on with value.
}
