\name{unit.test}
\alias{unit.test}
\alias{catch.out}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate unit tests}
\description{
Evaluates a assertion with n arguments provided by evaluating the generators. Stops or returns test cases.
}
\usage{
unit.test(assertion, generators = list(), sample.size = 10, stop = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{assertion}{}
  \item{generators}{}
  \item{sample.size}{}
  \item{stop}{}
	\item{...}{}
}
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\examples{

}
