% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{add_header}
\alias{add_header}
\title{Add header string to a list (as first element of the list). Helper of impose_accumulation_behavior.}
\usage{
add_header(header, alert_list)
}
\arguments{
\item{header}{string that will appear as header of the alert.}

\item{alert_list}{list to be formatted as an 'alert list'.}
}
\description{
Add header string to a list (as first element of the list). Helper of impose_accumulation_behavior.
}
