% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_lists_generators.R
\name{alert_generator.character}
\alias{alert_generator.character}
\title{S3 method for character class of alert_generator}
\usage{
\method{alert_generator}{character}(type, alert_message)
}
\arguments{
\item{type}{string equal to one of: 'error', 'message' and 'warning'. Determines the type of alert.}

\item{alert_message}{list of character vector or a single character vector that follow the cli
syntax. If list it is formatted following the scheme explained in format_cli_list.}
}
\description{
S3 method for character class of alert_generator
}
