% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generators.R
\name{alert_generator}
\alias{alert_generator}
\title{Function to signal errors, warnings and messages.}
\usage{
alert_generator(
  type,
  alert_message,
  eval_env,
  quickalert = TRUE,
  sign = TRUE,
  header = NULL,
  list_format = FALSE
)
}
\arguments{
\item{type}{String equal to one of: 'error', 'warning', 'message', 'accumulate_message'. Determines the type of alert.}

\item{alert_message}{Single character vector that follow the cli format.}

\item{eval_env}{Integer indicating the calling frame in which evaluates the alert in respect to where the alert is generated.
So it indicates the number of frames to look up in the calling stack.}

\item{quickalert}{Logical, whether to generate an alert with class "quickalert" or not (default TRUE).
The generation of "plain" alerts is useful when the checking functions are used inside other ones.
This allow the fact that only the main check raises quickalerts.}

\item{sign}{Logical, whether to add the sign to the first element of the alert (default TRUE).}

\item{header}{Character string to add at the beginning of the alert message (default NULL).}

\item{list_format}{Logical, whether to apply the list format style, which includes numbers of names
of the alert message elements to be represented in violet in a bullet list.}
}
\description{
Function to signal errors, warnings and messages.
}
