% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generator.R
\name{alert_generator}
\alias{alert_generator}
\title{Function to signal errors, warnings and messages.}
\usage{
alert_generator(
  type,
  alert_message,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  header = NULL,
  sign = TRUE,
  list_format = FALSE
)
}
\arguments{
\item{type}{String equal to one of: 'error', 'warning' or 'message'. Determines the type of alert.}

\item{alert_message}{Single character vector that follow the cli format. It possible to pass a list of strings to have automatically
a numbered or nominated list display (otherwise for character vectors use list_format = TRUE).}

\item{n_evaluation_frame}{Integer indicating the stack frame in which the alert glue expressions are evaluated in respect to where is generated.
So it indicates the number of frames to look down in the stack. By default is zero plus two.
The addition is used in order to start counting from zero from the frame of the function that uses this function.
Be aware that if the glue variables are not found in the pointed frame then the resolution falls under the
lexical scoping rules starting from the pointed frame.}

\item{quickalert}{Logical, whether to generate an alert with class "quickalert" or not (default TRUE).
The generation of "plain" alerts is useful when the checking functions are used inside other ones.
This allow that only the main check raises quickalerts.}

\item{header}{Character string to add at the beginning of the alert message (default NULL).}

\item{sign}{Logical, whether to add the sign to the first element of the alert (default TRUE).
Note that in case of errors the "!" sign is used as default by cli. This sign is changed with "X" if sign is TRUE,
but with sign = FALSE it cannot be suppressed.}

\item{list_format}{Logical, whether to apply the list format style, which includes numbers or names
of the alert message elements to be represented in violet in a bullet list.}
}
\value{
Raise a condition.
}
\description{
Function to signal errors, warnings and messages.
}
