% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.R
\name{anything}
\alias{anything}
\title{Any R object generator}
\usage{
anything(any_empty = TRUE, any_undefined = TRUE)
}
\arguments{
\item{any_empty}{Whether empty vectors or data frames should be allowed.}

\item{any_undefined}{Whether undefined values should be allowed.}
}
\value{
A \code{quickcheck_generator} object.
}
\description{
Generate any R object. This doesn't actually generate any possible object,
just the most common ones, namely atomic vectors, lists, data.frames,
tibbles, data.tables, and undefined values like \code{NA}, \code{NULL}, \code{Inf}, and
\code{NaN}.
}
\examples{
anything() |> show_example()
}
