% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{break_condition_message}
\alias{break_condition_message}
\title{Retrieve the condition message and breaks it in a character vector based on newline character.
This is helpful since quickalerts are defined using only the message argument of the cli-signaling functions.}
\usage{
break_condition_message(condition)
}
\arguments{
\item{condition}{condition object.}
}
\description{
Retrieve the condition message and breaks it in a character vector based on newline character.
This is helpful since quickalerts are defined using only the message argument of the cli-signaling functions.
}
