% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{break_condition_message}
\alias{break_condition_message}
\title{Retrieve the condition message and breaks it in a character vector based in newline character.
This is helpful since conditionMessage function returns both the message and the body of a condition object
as a single string (quickalerts have been defined using only the message argument, so all is a message strictly speaking).}
\usage{
break_condition_message(condition)
}
\arguments{
\item{condition}{condition object.}
}
\description{
Retrieve the condition message and breaks it in a character vector based in newline character.
This is helpful since conditionMessage function returns both the message and the body of a condition object
as a single string (quickalerts have been defined using only the message argument, so all is a message strictly speaking).
}
