% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_classes}
\alias{check_args_classes}
\title{Checks the classes of the selected arguments of a calling function.}
\usage{
check_args_classes(
  args,
  expected_classes,
  numeric_correspondence = NULL,
  null = FALSE,
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{args}{character vector reporting the arguments of the outer function to check.}

\item{expected_classes}{character vector with the expected classes (one for argument).}

\item{numeric_correspondence}{numeric vector that allows to recycle expected_types following
its order and the numbers specified in this vector.}

\item{null}{logical, whether args can be NULL or not in addiction to the "normal" expected type (default FALSE).}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of stack frame to look up for the evaluation of the glue expressions of the alert message.
The default value points to the frame of this function.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
Checks the classes of the selected arguments of a calling function.
}
