% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_classes}
\alias{check_args_classes}
\title{Checks the actual and expected classes of selected args of an outer function.}
\usage{
check_args_classes(args, expected_classes, numeric_correspondence = NULL)
}
\arguments{
\item{args}{character vector reporting the arguments of the outer calling function to check.}

\item{expected_classes}{character vector with the expected classes.}

\item{numeric_correspondence}{numeric vector that allows to recycle expected_types following
the order of expected_types and the number specified by this vector.}
}
\description{
Checks the actual and expected classes of selected args of an outer function.
}
