% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_primitive_types}
\alias{check_args_primitive_types}
\title{Checks the primitive types of the specified arguments of a calling function.}
\usage{
check_args_primitive_types(args, expected_types, numeric_correspondence = NULL)
}
\arguments{
\item{args}{character vector reporting the arguments of the outer function to check.}

\item{expected_types}{character vector with the expected types.}

\item{numeric_correspondence}{numeric vector that allows to recycle expected_types following
the order of expected_types and the number specified by this vector.}
}
\description{
Checks the primitive types of the specified arguments of a calling function.
}
