% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_primitive_types}
\alias{check_args_primitive_types}
\title{Checks the primitive types of the specified arguments of a calling function.}
\usage{
check_args_primitive_types(
  args,
  expected_types,
  numeric_correspondence = NULL,
  null = FALSE,
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{args}{character vector reporting the arguments of the outer function to check.}

\item{expected_types}{character vector with the expected types.}

\item{numeric_correspondence}{numeric vector that allows to recycle expected_types following
its order and the numbers specified in this vector.}

\item{null}{logical, whether args can be NULL or not in addiction to the "normal" expected type (default FALSE).}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.}

\item{header}{Character string to add at the beginning of the alert message.
If "default" the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message. The default value points to the frame below the one in which this function "exists"
(to not modify if the default alert is desired). So to point to the calling frame of this function
you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
Checks the primitive types of the specified arguments of a calling function.
}
