% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_columns_copresence}
\alias{check_columns_copresence}
\title{Check duplicated column names in two dataframes}
\usage{
check_columns_copresence(
  df1,
  df2,
  columns = NULL,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{columns}{Character vector of columns to consider. If NULL all columns are considered (default NULL).}

\item{df1_arg}{String specifying how to address first df in the raised messages (default "df1").}

\item{df2_arg}{String specifying how to address second df in the raised messages (default "df2").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{String to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
Check duplicated column names in two dataframes
}
