% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_key}
\alias{check_columns_key}
\title{Checks if the specified dataframe columns are suitable as keys of only unique values.}
\usage{
check_columns_key(
  df,
  columns,
  na.rm = TRUE,
  raise = "error",
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{Character string reporting the column/s name.}

\item{na.rm}{logical (default TRUE), indicating if NA must be excluded prior evaluation.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{string; if equal to 'default' the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message. The default value points to the frame below the one created by this function
(to not modify if the default alert is desired). So to point to the calling frame of this function
you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL.
}
\description{
Checks if the specified dataframe columns are suitable as keys of only unique values.
}
