% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_key}
\alias{check_columns_key}
\title{Checks if the specified dataframe columns are suitable as keys of only unique values.}
\usage{
check_columns_key(
  df,
  keys,
  raise = "error",
  alert_message = NULL,
  na.rm = TRUE,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{keys}{character vector reporting the names of the columns to test as key.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{na.rm}{logical (default TRUE), indicating if NA must be excluded prior computations.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired).
So it's easy to points to upper frames (as well as to below frames but is not recommended).}
}
\value{
NULL.
}
\description{
Checks if the specified dataframe columns are suitable as keys of only unique values.
}
