% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_key}
\alias{check_columns_key}
\title{Checks if the specified dataframe columns are suitable as keys of only unique values.}
\usage{
check_columns_key(
  df,
  columns,
  na.rm = TRUE,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2,
  quickalert = TRUE
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{Character string reporting the column/s name.}

\item{na.rm}{logical (default TRUE), indicating if NA must be excluded prior evaluation.}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be ignored
most of times because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired). So to point to the calling frame of this function you have to set 4.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert".}
}
\value{
invisible NULL.
}
\description{
Checks if the specified dataframe columns are suitable as keys of only unique values.
}
\details{
Is suggested to not set a custom alert message. This is for the internal working of the function which
will trigger the custom message for every column that fails the checking.
}
