% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_levels}
\alias{check_columns_levels}
\title{Checks the presence of the specified values in the selected column}
\usage{
check_columns_levels(
  df,
  columns,
  col_levels,
  raise = "error",
  alert_message = NULL
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{character vector of columns names to check}

\item{col_levels}{list of character vector reporting the expected levels for each column
specified in columns. The element of the list must be nominated according to the columns values.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}
}
\value{
invisible NULL
}
\description{
Checks the presence of the specified values in the selected column
}
