% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_levels}
\alias{check_columns_levels}
\title{Checks the presence of the specified values in the selected columns.}
\usage{
check_columns_levels(
  df,
  columns,
  col_levels,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{character vector of columns names to check}

\item{col_levels}{list of character vector reporting the expected levels for each column
specified in columns. The element of the list must be nominated according to the columns values.}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be ignored
most of times because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired). So to point to the calling frame of this function you have to set 4.}
}
\description{
Checks the presence of the specified values in the selected columns.
}
\details{
Is suggested to not set a custom alert message. This is for the internal working of the function which
will trigger the custom message for every column that fails the checking.
}
