% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_levels}
\alias{check_columns_levels}
\title{Checks the presence of the specified values in the selected columns.}
\usage{
check_columns_levels(
  df,
  columns,
  col_levels,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{character vector of columns names to check}

\item{col_levels}{list of character vector reporting the expected levels for each column
specified in columns. The element of the list must be nominated according to the columns values.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired).
So it's easy to points to upper frames (as well as to below frames but is not recommended).}
}
\description{
Checks the presence of the specified values in the selected columns.
}
