% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_na}
\alias{check_columns_na}
\title{Checks the presence of NAs in the specified columns}
\usage{
check_columns_na(
  df,
  columns,
  raise = "error",
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{Character string reporting the column/s name.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Default NULL, in this case a standard
message with the appropriate bullet sign is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{string; if equal to 'default' the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message. The default value points to the frame below the one in which this function "exists"
(to not modify if the default alert is desired). So to point to the calling frame of this function
you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks the presence of NAs in the specified columns
}
