% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R, R/make_test_function.R
\name{check_columns_number}
\alias{check_columns_number}
\alias{test_columns_number}
\title{Check the number of columns of a dataframe}
\usage{
check_columns_number(
  df,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  df_arg = "df",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)

test_columns_number(
  df,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  df_arg = "df",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{exact_len}{Integer indicating the exact expected number of columns (default NULL).}

\item{min_len}{Integer indicating the minimum expected number of columns (default NULL).}

\item{max_len}{Integer indicating the maximum expected number of columns (default NULL).}

\item{df_arg}{String specifying how to address df in the raised messages (default "df").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to the frame above this function frame. So to point to the frame below this function frame you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Check the number of columns of a dataframe
}
