% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_predicate}
\alias{check_columns_predicate}
\title{Checks if all columns satisfy the predicate.}
\usage{
check_columns_predicate(
  df,
  predicate,
  inverse = FALSE,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{predicate}{function that works on vectors and return a single logical value.}

\item{inverse}{logical, whether to invert the check direction in the sense that the predicate
must be not satisfied for all columns (default FALSE).}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be most of
times ignored because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame (to not modify
if the default alert is desired). So to point to the calling frame of this function you have to set 4.}
}
\description{
Checks if all columns satisfy the predicate.
}
