% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_presence}
\alias{check_columns_presence}
\title{Checks the presence of one or multiple columns in a dataframe.}
\usage{
check_columns_presence(
  df,
  columns,
  df_arg = "df",
  raise = "error",
  warn_message = NULL
)
}
\arguments{
\item{df}{dataframe.}

\item{columns}{character string reporting the column/s name.
@param df_arg string specifying how to address df in the raised messages (default "df").}

\item{raise}{character string equal to "error" or "warning" (default error).}

\item{warn_message}{string reporting the warning or message based on raise value
(default NULL, in this case the error message is used with the correct sign).}
}
\value{
NULL or side effects.
}
\description{
Checks the presence of a column in a dataframe base on its name and raises error, warning or a message.
The error message is standardized because always informative, the warning and message can be personalized,
following the cli syntax (it is passed to cli_bullets).
If NULL the message displayed for errors is used with the "correct bullet sign".
}
