% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_columns_presence}
\alias{check_columns_presence}
\title{Checks the presence of one or multiple columns in a dataframe}
\usage{
check_columns_presence(
  df,
  columns,
  df_arg = "df",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{columns}{Character string reporting the column/s name.}

\item{df_arg}{String specifying how to address df in the raised messages (default "df").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{Numeric, defines the number of calling frame to look down for the evaluation of the glue expressions of the alert message.
The default value points to the frame above the one of this function (to not modify if the default alert is desired).
So to point to the frame below this function frame you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
NULL.
}
\description{
Checks the presence of one or multiple columns in a dataframe
}
