% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vec.R
\name{check_empty_vec}
\alias{check_empty_vec}
\title{Check the emptiness of a vector}
\usage{
check_empty_vec(
  vec,
  na = TRUE,
  empty_string = TRUE,
  len = TRUE,
  null = TRUE,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{na}{Boolean, indicating whether to perform the check for NAs (default TRUE).}

\item{empty_string}{Boolean, indicating whether to perform the check for empty string "" (default TRUE).}

\item{len}{Boolean, indicating whether to perform the check for 0 length (default TRUE).}

\item{null}{Boolean, indicating whether to perform the check for NULL value (default TRUE).}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks the emptiness of a vector and or the presence of "empty" values in it.
Empty here is declined in its broader meaning indicating NAs, "", NULL and vector of length zero.
This interpretation can be fine tuned using the parameters function.
}
