% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_empty_vec}
\alias{check_empty_vec}
\title{Checks the presence of "empty" values in one vector. Empty here is declined in its
broader meaning indicating NAs, "", NULL and vector of length zero.}
\usage{
check_empty_vec(vec, vec_arg = "vec", raise = "error", alert_message = NULL)
}
\arguments{
\item{vec}{vector to check}

\item{vec_arg}{string indicating how to address vec in the alert message (default 'vec')}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}
}
\value{
invisible NULL
}
\description{
Checks the presence of "empty" values in one vector. Empty here is declined in its
broader meaning indicating NAs, "", NULL and vector of length zero.
}
