% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_empty_vec}
\alias{check_empty_vec}
\title{Checks the presence of "empty" values in one vector. Empty here is declined in its
broader meaning indicating NAs, "", NULL and vector of length zero.}
\usage{
check_empty_vec(
  vec,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be most of
times ignored because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the function frame. So it's possible to simply points
to upper frames (as well as to below frames but is not recommended).}

\item{...}{not of direct use.}
}
\value{
invisible NULL
}
\description{
Checks the presence of "empty" values in one vector. Empty here is declined in its
broader meaning indicating NAs, "", NULL and vector of length zero.
}
