% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_key}
\alias{check_key}
\title{Checks if the specified column of the dataframe is suitable as key with only unique values.}
\usage{
check_key(
  df,
  key,
  raise = "error",
  alert_message = NULL,
  na.rm = TRUE,
  n.evaluation_frame = 2
)
}
\arguments{
\item{df}{Dataframe passed in the outer function.}

\item{key}{character reporting the column name.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{String reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}

\item{na.rm}{logical (default TRUE), indicating if NA must be excluded prior computations.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the frame below the function frame. So it's possible to simply points
to upper frames (as well as to below frames but is not recommended).}
}
\value{
invisible NULL.
}
\description{
Checks if the specified column of the dataframe is suitable as key with only unique values.
}
