% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_df.R
\name{check_key}
\alias{check_key}
\title{Checks if the specified column of the dataframe is suitable as key with only unique values.}
\usage{
check_key(df, key, raise = "error", alert_message = NULL)
}
\arguments{
\item{df}{dataframe.}

\item{key}{character reporting the column name.}

\item{raise}{character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{string reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}
}
\value{
NULL.
}
\description{
Checks if the specified column of the dataframe is suitable as key with only unique values.
}
