% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_vecs.R, R/make_test_function.R
\name{check_length_vecs}
\alias{check_length_vecs}
\alias{test_lenght_vecs}
\title{Check whether two vectors have the same length}
\usage{
check_length_vecs(
  vec1,
  vec2,
  na_rm = FALSE,
  unique = FALSE,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n_evaluation_frame = 0,
  ...
)

test_lenght_vecs(
  vec1,
  vec2,
  na_rm = FALSE,
  unique = FALSE,
  vec1_arg = "vec1",
  vec2_arg = "vec2",
  raise = "error",
  alert_message = NULL,
  quickalert = TRUE,
  n_evaluation_frame = 0,
  ...
)
}
\arguments{
\item{vec1}{First vector.}

\item{vec2}{Second vector.}

\item{na_rm}{Boolean indicating if NA must be excluded prior checking (default FALSE).}

\item{unique}{Boolean indicating whether to perform the check only on unique values (default FALSE).}

\item{vec1_arg}{String indicating how to address vec1 in the raised message (default 'vec1').}

\item{vec2_arg}{String indicating how to address vec2 in the raised message (default 'vec2)'.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\value{
Depending on the function prefix: the "check" function returns the condition otherwise NULL invisibly,
the "test" function returns TRUE if the condition would be raised and FALSE otherwise.
}
\description{
Check whether two vectors have the same length
}
