% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_nrow}
\alias{check_nrow}
\title{Checks equal number of rows between two dataframes}
\usage{
check_nrow(
  df1,
  df2,
  df1_arg = "df1",
  df2_arg = "df2",
  raise = "error",
  alert_message = NULL
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{df1_arg}{string specifying how to address df1 in the raised messages (default "df1").}

\item{df2_arg}{string specifying how to address df2 in the raised messages (default "df2").}

\item{raise}{character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{string reporting the alert message. Its formatted by cli_bullets function.
Default NULL, in this case a standard message with the appropriate bullet sign is used.}
}
\value{
NULL or side effects
}
\description{
Checks equal number of rows between two dataframes
}
