% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_nrow_dfs}
\alias{check_nrow_dfs}
\title{Checks equal number of rows between two dataframes}
\usage{
check_nrow_dfs(
  df1,
  df2,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message. The default value points to the frame below the one created by this function
(to not modify if the default alert is desired). So to point to the calling frame of this function
you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
Checks equal number of rows between two dataframes
}
