% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_number_values}
\alias{check_number_values}
\title{Checks if the number of unique values of a vector is equal to the expected.}
\usage{
check_number_values(
  vec,
  expected_number_levels,
  vec_arg = "vec",
  na.rm = TRUE,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{expected_number_levels}{numeric indicating the expected number of unique values for vec.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{na.rm}{logical (default TRUE), indicating if NA must be excluded prior computations.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation of the alert message.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks if the number of unique values of a vector is equal to the expected.
}
