% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_numeric_args}
\alias{check_numeric_args}
\title{Checks the numeric nature of selected arguments of a calling function.}
\usage{
check_numeric_args(
  args,
  null = FALSE,
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{args}{character vector reporting the arguments of the outer function to check.}

\item{null}{logical, whether args can be NULL or not in addiction to the "normal" expected type (default FALSE).}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
Allow to easily check for numeric arguments since is not easy to do that relying on classes or primitive types.
}
