% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_list.R
\name{check_predicate_list}
\alias{check_predicate_list}
\title{Checks all the elements of the list through a predicate function.}
\usage{
check_predicate_list(
  x,
  predicate,
  flatten = TRUE,
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{x}{list.}

\item{predicate}{A predicate function, usually an "is.something" function.}

\item{flatten}{logical, whether to flat out the list before doing the check (default TRUE).}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message.
If "default" the default header is used, otherwise the string passed in.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL.
}
\description{
The function doesn't perform any check on the argument provided in predicate.
Therefore the correctness of the provided function fall on the user.
}
\details{
The alert will not points the eventual elements in error if no or some elements names are missing,
but instead it will raise a more general alert.
}
