% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross_dfs.R
\name{check_presence_dfs}
\alias{check_presence_dfs}
\title{Perform an ordered cross-checking between the values of two columns of two dataframes.}
\usage{
check_presence_dfs(
  df1,
  df2,
  col,
  direction = "first_in_second",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{df1}{first dataframe.}

\item{df2}{second dataframe.}

\item{col}{character vector referring the name of the columns to consider. If of length one
the single name is assumed for both dataframes. If of length two the first name is assumed
for the first dataframe and the other for the second.}

\item{direction}{string equal to one of 'first_in_second', 'second_in_first' or 'bidirectional'.
Set the direction of the comparison.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message. The default value points to the frame below the one created by this function
(to not modify if the default alert is desired). So to point to the calling frame of this function
you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\description{
The function allows to check the presence of all values of the selected column of one dataframe
in the other. If direction equal 'bidirectional' a bidirectional check is performed.
}
