% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_presence_values}
\alias{check_presence_values}
\title{Checks the presence of the specified values in a vector.}
\usage{
check_presence_values(
  vec,
  values,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 2,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{values}{character vector of values searched in vec.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning", "message" or "accumulate_message"
(default error). Set the type of alert that is created. Note: 'accumulate_message' can be ignored
most of times because it is used internally in combination with impose_accumulate_behavior.
Either the case it raises a message without the "i" bullet.}

\item{alert_message}{String reporting the alert message. Default NULL, in this case a standard
message with the appropriate bullet sign is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run.
The default value points to the function frame. So it's possible to simply points
to upper frames (as well as to below frames but is not recommended).}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks the presence of the specified values in a vector.
}
