% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vector.R
\name{check_presence_values}
\alias{check_presence_values}
\title{Checks the presence of the specified values in a vector.}
\usage{
check_presence_values(
  vec,
  values,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{values}{character vector of values searched in vec.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation of the alert message.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks the presence of the specified values in a vector.
}
