% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vec.R
\name{check_sorted_vec}
\alias{check_sorted_vec}
\title{Checks if a vector is sorted or not.}
\usage{
check_sorted_vec(
  vec,
  vec_arg = "vec",
  decreasing = FALSE,
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{decreasing}{Logical indicating whether the expect sorted order is decreasing or not (default FALSE).}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Checks if a vector is sorted or not.
}
