% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_lists_generators.R
\name{cli_features}
\alias{cli_features}
\alias{generate_cli_alert_list}
\title{Set of functions that manipulate cli functions and facilitate apllying its syntax.
Generates a list of main cli alert functions with informative names.}
\usage{
generate_cli_alert_list()
}
\description{
Set of functions that manipulate cli functions and facilitate apllying its syntax.
Generates a list of main cli alert functions with informative names.
}
