% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant}
\alias{constant}
\title{Generate the same value every time}
\usage{
constant(a)
}
\arguments{
\item{a}{Any R object}
}
\value{
A \code{hedgehog.internal.gen} generator
}
\description{
Generate the same value every time
}
\examples{
constant(NULL) |> show_example()
}
