% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_all.R
\name{for_all}
\alias{for_all}
\title{Test properties of a function}
\usage{
for_all(
  ...,
  property,
  tests = get_tests(),
  size = get_size(),
  shrinks = get_shrinks(),
  discards = get_discards()
)
}
\arguments{
\item{...}{Dots}

\item{property}{A function which takes a value from from
the generator and calls an expectation on it.}

\item{tests}{The number of tests to run.}

\item{size}{The maximum length of the generators.}

\item{shrinks}{The maximum number of shrinks to run when
shrinking a value to find the smallest counterexample.}

\item{discards}{The maximum number of discards to permit
when running the property.}
}
\value{
TRUE if tests pass
}
\description{
Test properties of a function
}
\examples{
library(testthat)
for_all(
  a = numeric_(),
  b = numeric_(),
  property = \(a, b) expect_equal(a + b, b + a)
)
for_all(
  x = any_vector(),
  property = \(x) rev(x) |> rev() |> expect_equal(x)
)
}
