% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_generator.R
\name{generate_header}
\alias{generate_header}
\title{Constructs a default header message if equal to "default".}
\usage{
generate_header(header, default_header)
}
\arguments{
\item{header}{takes in the header argument of the checking function.}

\item{default_header}{string that is used as default header.}
}
\description{
Constructs a default header message if equal to "default".
}
