% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_check.R
\name{generic_check}
\alias{generic_check}
\title{Generic quickcheck constructor.}
\usage{
generic_check(
  expr,
  raise = "error",
  alert_message = NULL,
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{expr}{expression to test. Must return a single boolean value.}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n.evaluation_frame}{Numeric, defines the number of calling frame to look up for the evaluation of the glue expressions of the alert message.
The default value points to the frame below the one of this function (to not modify if the default alert is desired).
So to point to the upper frame of this function you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
Invisible NULL
}
\description{
Allows to test generic hypothesis while having all the additional features of quickcheck.
The alert is raise if the provided expression evaluates to FALSE.
}
