% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_accumulation_behavior}
\alias{impose_accumulation_behavior}
\title{Allow to accumulate in list non-error alert of checking functions}
\usage{
impose_accumulation_behavior(
  expr,
  raise = "error",
  alert_message = NULL,
  header = NULL,
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{expr}{check function call. It's important that the checking function raises warnings or messages
since errors cannot be muffled (they will cause unexpected behavior).}

\item{raise}{type of the accumulated final alert if any.}

\item{alert_message}{String or list reporting the alert message (by default the function build a list).}

\item{header}{string to add as the header of the accumulated alert list.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run. The default points
to this function frame.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Imposes an accumulation behavior for a checking function in a loop scenario, in which
the alert raised by the checking function are stored in a list and then displayed.
Works with message or warning conditions. It does not work with error since they stop the loop execution.
}
