% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_accumulation_behavior}
\alias{impose_accumulation_behavior}
\title{Accumulate and then raise non-error quickalerts}
\usage{
impose_accumulation_behavior(
  expr,
  raise = "error",
  alert_message = NULL,
  header = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{expr}{check function call. It's important that the checking function raises warnings or messages and not errors.}

\item{raise}{type of the accumulated final alert if any.}

\item{alert_message}{String or list reporting the alert message (by default the function build a list).}

\item{header}{string to add as the header of the accumulated alert list.}

\item{n_evaluation_frame}{Intergerish, defines the number of stack frame to look down for the evaluation
of the glue expressions of the alert message. The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert" (default TRUE).}

\item{...}{To pass additional parameters to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Imposes an accumulation behavior in which quickalerts are stored in a list, which is used to raise a condition.
Works only with messages and warnings (not errors).
}
