% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_additional_alert}
\alias{impose_additional_alert}
\title{Adds a message at the start/end of raised quickalert messages.}
\usage{
impose_additional_alert(
  expr,
  message,
  margin = 1,
  raise = "error",
  n.evaluation_frame = 0,
  quickalert = TRUE
)
}
\arguments{
\item{expr}{check function call. It's important that the checking function raises warnings or messages but not errors.}

\item{message}{additional message to be added to the alert.}

\item{margin}{numeric equal to 1 or 2, indicating where to add the additional message (start or end respectively).}

\item{raise}{type of the accumulated final alert if any.}

\item{n.evaluation_frame}{numeric, defines the number of stack frame to look up for the evaluation
of the glue expressions of the alert message. The default points to this function frame.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert" (default TRUE).}
}
\value{
invisible NULL
}
\description{
Adds a message at the start/end of raised quickalert messages.
}
\details{
The alert sign of the resulting message (if any) is decided by the type of alert raised by the checking function in expr.
This has been chosen to prevent that one must set sign argument to FALSE for the checking function all times.
}
