% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_additional_alert}
\alias{impose_additional_alert}
\title{Allow the add a message at the start/end of raised quickalert messages}
\usage{
impose_additional_alert(
  expr,
  message,
  margin = 1,
  raise = "error",
  n.evaluation_frame = 2,
  quickalert = TRUE
)
}
\arguments{
\item{expr}{check function call.}

\item{message}{additional message to be added to the alert.}

\item{margin}{numeric equal to 1 or 2, indicating where to add the additional message (start or end respectively).}

\item{raise}{type of the accumulated final alert if any.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run. The default points
to this function frame.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert".}
}
\value{
invisible NULL
}
\description{
Allow the add a message at the start/end of raised quickalert messages
}
