% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_logical_behavior}
\alias{impose_logical_behavior}
\title{Alter the behavior of checking conditions from raising alerts to return boolean values.}
\usage{
impose_logical_behavior(expr)
}
\arguments{
\item{expr}{check function call.}
}
\value{
A single logical value.
}
\description{
This function works as a wrapper of the checking functions and impose them a
'logical' behavior. This means that they will not raise anymore alerts but instead
they will return boolean values. In greater detail if a 'quickalert' class
condition is hit, then the function return TRUE otherwise FALSE.
}
