% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_loop_behavior}
\alias{impose_loop_behavior}
\title{Perform the check on all the elements of a list generating a single summarized alert}
\usage{
impose_loop_behavior(
  x,
  check_func,
  check_arg_list = list(),
  element_nameroot = "element",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{x}{List which elements will be checked by the check_func.}

\item{check_func}{Checking function that will be used on every x element (this will be passed as first argument).}

\item{check_arg_list}{List of arguments to pass to check_func.
Note that this list must not contain 'quickalert-oriented' arguments (raise, header, quickalert and sign),
neither the objects on which the checks are done, since these are automatically set.}

\item{element_nameroot}{String reporting the base name to use for the list elements with an empty name (default "element").}

\item{raise}{type of the accumulated final alert if any.}

\item{alert_message}{String or list reporting the alert message (by default the function build a list).}

\item{header}{string to add as the header of the accumulated alert list.}

\item{n_evaluation_frame}{Intergerish, defines the number of stack frame to look down for the evaluation
of the glue expressions of the alert message. The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert is of class "quickalert" (default TRUE).}

\item{...}{To pass additional parameters to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Iterates the checking function on all the elements of the list generating alerts
that pinpoint the list names/positions for which the checking function failed.
In greater detail, the names are used if not-empty otherwise the positions.
}
