% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_wrappers.R
\name{impose_loop_behavior}
\alias{impose_loop_behavior}
\title{Perform the same check to all the elements of a list generating a single summarized alert.}
\usage{
impose_loop_behavior(
  x,
  check_func,
  check_arg_list = list(),
  element_nameroot = "element",
  raise = "error",
  alert_message = NULL,
  header = "default",
  n.evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{x}{List which elements will be checked by the check_func.}

\item{check_func}{Checking function that will be used on every x element (this will be passed as first argument).}

\item{check_arg_list}{List of arguments to pass to check_func. Importantly the first checking function argument
as well as 'alert-oriented' arguments (raise, header, quickalert and sign) are already set and therefore must NOT be specified.}

\item{element_nameroot}{String reporting the basename to use for the list elements names without a name in case
they must be pointed in the alert message (default "element").}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error).
Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{header}{Character string to add at the beginning of the alert message.
If "default" the default header is used, otherwise the string passed in.}

\item{n.evaluation_frame}{numeric, defines the number of calling frame to look up for the evaluation
of the alert message in respect to where the function calling the alert is run. The default points
to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass additional argument to alert_generator function.}
}
\value{
invisible NULL
}
\description{
Iterate the checking function on all the elements of the list and generate informative alert messages,
that pinpoint the element names/positions for which the checking function failed.
}
