% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_single_vec.R
\name{internal_check_length_vec}
\alias{internal_check_length_vec}
\title{Internal check of check_length_vec}
\usage{
internal_check_length_vec(
  vec,
  exact_len = NULL,
  min_len = NULL,
  max_len = NULL,
  na_rm = TRUE,
  unique = FALSE,
  vec_arg = "vec",
  raise = "error",
  alert_message = NULL,
  n_evaluation_frame = 0,
  quickalert = TRUE,
  ...
)
}
\arguments{
\item{vec}{Vector to check.}

\item{exact_len}{Integer indicating the exact expected vector length (default NULL).}

\item{min_len}{Integer indicating the minimum expected vector length (default NULL).}

\item{max_len}{Integer indicating the maximum expected vector length (default NULL).}

\item{na_rm}{Boolean indicating if NA must be excluded prior checking (default TRUE).}

\item{unique}{Boolean indicating whether to perform the check only on unique values (default FALSE).}

\item{vec_arg}{String indicating how to address vec in the alert message (default 'vec').}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default value (0) points to the frame above this function frame. So to point to the frame below this function frame you have to set 2.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Internal check of check_length_vec
}
