% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_vec.R
\name{is_empty_vec}
\alias{is_empty_vec}
\title{Checks the presence of null values in a vector in a broader sense.}
\usage{
is_empty_vec(vec, null = TRUE, len = TRUE, na = TRUE, empty_string = TRUE)
}
\arguments{
\item{vec}{vector to test.}

\item{null}{Boolean, indicating whether to perform the check for NULL value (default TRUE).}

\item{len}{Boolean, indicating whether to perform the check for 0 length (default TRUE).}

\item{na}{Boolean, indicating whether to perform the check for NAs (default TRUE).}

\item{empty_string}{Boolean, indicating whether to perform the check for empty string "" (default TRUE).}
}
\value{
A single boolean, FALSE even if one value is empty, TRUE otherwise.
}
\description{
Checks the presence of null values in a vector in a broader sense.
}
