% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is_empty_vec}
\alias{is_empty_vec}
\title{Checks the presence of null values in a vector in a broader sense.
Return a single boolean (FALSE even if one value is empty, TRUE otherwise).}
\usage{
is_empty_vec(vec)
}
\arguments{
\item{vec}{vector to test.}
}
\description{
Checks the presence of null values in a vector in a broader sense.
Return a single boolean (FALSE even if one value is empty, TRUE otherwise).
}
