% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_}
\alias{list_}
\title{List generator}
\usage{
list_(...)
}
\arguments{
\item{...}{A set of named or unnamed generators.}
}
\value{
A \code{quickcheck_generator} object.
}
\description{
Generate lists with contents corresponding to the values generated by the
input generators.
}
\examples{
list_(integer_(), logical_()) \%>\% show_example()
list_(a = any_vector(), b = any_vector()) \%>\% show_example()
}
