\name{quickcheck-package}
\alias{quickcheck-package}
\alias{quickcheck}
\docType{package}
\title{
Randomized testing
}
\description{
Supports the generation of randomized tests by providing functions to define tests and reproduce errors and a collection of data generators for all the most common types.
}
\details{
See \code{packageDescription("quickcheck")}
}
\author{
Revolution Analytics <rhadoop@revolutionanalytics.com>
}

\keyword{testing }
