\name{quickcheck-package}
\alias{quickcheck-package}
\alias{quickcheck}
\docType{package}
\title{
Randomized unit testing
}
\description{
Supports the generatio of randomized unit tests by defining the \code{\link{unit.test}} function and a collection of meta-genarators for data of all types
}
\details{
\tabular{ll}{
Package: \tab quickcheck\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-15\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Revolution Analytics <rhadoop@revolutionanalytics.com>
}
\references{
}
\keyword{testing }
\seealso{

}
\examples{
~~ simple examples of the most important functions ~~
}
