% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasversal_internals.R
\name{raise_predicate_error}
\alias{raise_predicate_error}
\title{Helper to raise alert for check_predicate_* funcs.}
\usage{
raise_predicate_error(
  log_empty_names,
  errors,
  string_object,
  inverse,
  raise,
  alert_message,
  n_evaluation_frame,
  quickalert,
  header,
  ...
)
}
\arguments{
\item{log_empty_names}{Boolean indicating if the object on which the predicate is checked has empty names.}

\item{errors}{Vector with the names of the error elements or 1.}

\item{string_object}{String with the name of the object to use inside the alerts.}

\item{inverse}{Boolean, whether to invert the check direction in the sense that the predicate must be not satisfied for all the elements (default FALSE).}

\item{raise}{Character string equal to one of "error", "warning" or "message" (default error). Set the type of alert that is created.}

\item{alert_message}{Character vector reporting the alert message. Default NULL, in this case a standard message is used.
It's also possible to pass a list of strings that is displayed as a nominated or numbered list.}

\item{n_evaluation_frame}{Numeric, defines the number of stack frame to look down for the evaluation of the glue expressions of the alert message.
The default (0) points to this function frame.}

\item{quickalert}{logical, whether the raised alert has to be of class "quickalert".}

\item{header}{Character string to add at the beginning of the alert message. If "default" the default header is used, otherwise the string passed in.}

\item{...}{To pass the additional parameters sign, list_format and header (if not in the formals parameters).}
}
\description{
Helper to raise alert for check_predicate_* funcs.
}
