% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{show_example}
\alias{show_example}
\title{Show an example output of a generator}
\usage{
show_example(generator)
}
\arguments{
\item{generator}{A generator}
}
\value{
An example output produced by the generator.
}
\description{
Show an example output of a generator
}
\examples{
logical_() |> show_example()
}
